Handle missing-module-docstring correctly.
missing-module-docstring comment needs to be on the first line by itself.
If there is another command on this line, pylint ignores the comment.

Also in this commit:

Removed bad-whitespace in test since it was removed in pylint 2.6.

Use multiprocessing to run pylint in test. The privious implementation did
not work correctly on the second invocation of pylint.