Preserve comments meant for other non-pylint tools (#2)
When running `pylint-silent reset`, the current behavior is to strip any
comments that can come *after* `# pylint: disable=`. This can cause
breakages for those other tools (including `mypy`, `pants`, `black`,
`isort`, others).

Of course, you *can* make sure that every `# pylint: disable=...` comment
is the last one on its line, but it's much nicer if we can preserve any
other comments that follow.

Change the behavior of `reset` so that it does not strip comments meant
for other tools.

This approach intentionally avoids using `re` (in keeping with the style
of existing code). It should impose a very, very mild hit to performance.